[ALPS02315831] EMI: support 4GB mode
[Detail]change to unsigned long long for 4GB address, and not 4GB should have a offset

[Solution]change to unsigned long long for 4GB address, and not 4GB should have a offset

[Feature] 4GB_DRAM

MTK-Commit-Id: aac4de74e34e1472d3637a7c9292356e9d797bb5

CR-Id: ALPS02315831

+++ drivers/misc/mediatek/emi_mpu/emi_reg_rw.c
@@ -58,8 +58,8 @@ unsigned int mt_emi_reg_read(unsigned int offset)
 	return 0;
 }
 
-int mt_emi_mpu_set_region_protection(unsigned int start,
-				     unsigned int end,
+int mt_emi_mpu_set_region_protection(unsigned long long start,
+				     unsigned long long end,
 				     unsigned int region_permission)
 {
 #if defined(CONFIG_ARM_PSCI) || defined(CONFIG_MTK_PSCI)
